Made a dedicated file for netowrk errors
